+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## master
+
+* [minimagick] Add `:loader` option for explicitly setting input file type (@janko)
+
 ## 1.9.3 (2019-08-11)
 
 * [vips] Use integer precision when sharpening for better quality (@metaskills)